refactor: move all build in to one file, and remove constants.ts file
