+++ packages/g6/src/animations/executor.ts
@@ -2,8 +2,14 @@ import type { DisplayObject, IAnimation } from '@antv/g';
 import { isString, upperFirst } from '@antv/util';
 import { getPlugin } from '../registry';
 import { createAnimationsProxy, executeAnimation, inferDefaultValue, preprocessKeyframes } from '../utils/animation';
-import { DEFAULT_ANIMATION_OPTIONS } from './constants';
 import type { AnimationExecutor } from './types';
+
+const DEFAULT_ANIMATION_OPTIONS: KeyframeAnimationOptions = {
+  duration: 1000,
+  easing: 'cubic-bezier(0.250, 0.460, 0.450, 0.940)',
+  iterations: 1,
+  fill: 'both',
+};
 
 /**
  * <zh/> 动画 Spec 执行器
