+++ packages/g6/src/animations/index.ts
@@ -1,2 +1,17 @@
-export { BUILT_IN_ANIMATIONS } from './constants';
 export { executor } from './executor';
+
+/**
+ * <zh/> 内置的动画元素。
+ * <en/> Built-in animations.
+ */
+export const fade = [
+  {
+    fields: ['opacity'],
+  },
+];
+
+export const translate = [
+  {
+    fields: ['x', 'y'],
+  },
+];
