+++ packages/g6/src/layouts/index.ts
@@ -1,1 +1,15 @@
-export { BUILT_IN_LAYOUTS } from './constants';
+export { compactBox, dendrogram, indented, mindmap } from '@antv/hierarchy';
+export {
+  CircularLayout,
+  ComboCombinedLayout,
+  ConcentricLayout,
+  D3ForceLayout,
+  DagreLayout,
+  ForceAtlas2Layout,
+  ForceLayout,
+  FruchtermanLayout,
+  GridLayout,
+  MDSLayout,
+  RadialLayout,
+  RandomLayout,
+} from '@antv/layout';
