--- packages/g6/src/registry/constants.ts
@@ -1,17 +1,0 @@
-import { BUILT_IN_ANIMATIONS } from '../animations';
-import { BUILT_IN_EDGES, BUILT_IN_NODES } from '../elements';
-import { BUILT_IN_LAYOUTS } from '../layouts';
-import { BUILT_IN_PALETTES } from '../palettes';
-import { BUILT_IN_THEMES } from '../themes';
-
-export const BUILT_IN_PLUGINS = {
-  animation: BUILT_IN_ANIMATIONS,
-  behavior: {},
-  combo: {},
-  edge: BUILT_IN_EDGES,
-  layout: BUILT_IN_LAYOUTS,
-  node: BUILT_IN_NODES,
-  palette: BUILT_IN_PALETTES,
-  theme: BUILT_IN_THEMES,
-  widget: {},
-};
