+++ packages/g6/src/themes/dark.ts
@@ -1,6 +1,6 @@
-import { Theme } from './types';
+import type { Theme } from './types';
 
-export const DARK_THEME: Theme = {
+export const dark: Theme = {
   node: {
     style: {
       fill: '#444',
