+++ packages/g6/src/themes/index.ts
@@ -1,12 +1,2 @@
-import { DARK_THEME } from './dark';
-import { LIGHT_THEME } from './light';
-
-/**
- * <zh/> 内置主题
- *
- * <en/> Built-in themes
- */
-export const BUILT_IN_THEMES = {
-  light: LIGHT_THEME,
-  dark: DARK_THEME,
-};
+export { dark } from './dark';
+export { light } from './light';
